Avoid showing ``cms_pages`` twice in the product admin
If the user does not explicitly define the fields or fieldsets for their
product admin class, they will end up with the M2M widget for ``cms_pages``
being displayed twice. This does not occur in our demo shops, because there we
explicitly set ``fieldsets``. This commit changes the behavior of
CMSPageAsCategoryMixin so it removes the existing ``cms_pages`` field.